\name{hydroEFS-package}
\alias{hydroEFS-package}
\alias{hydroEFS}
\docType{package}
\title{
Hydrological Event Finding and Summarising (hydroEFS).
}
\description{
This package allows users to find events, based on the conditions of the ISCO 3700 storm sampling program. It also provides the ability to obtain summarises of before and during events.
}
\details{
\tabular{ll}{
Package: \tab hydroEFS\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2011-06-06\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Jason Lessels

Maintainer: Jason Lessels <jason.lessels@sydney.edu.au>
}
